procmail: Change a position of newline in original/procmailrc.5.
The original document had a line which begins with a single quote.
Precisely speaking it is not a valid syntax in a roff format and
it results in an error when processed by po4a-gettextize.
This commit moves a newline to avoid the error without changing
the content of that sentense.