Temporary commit of PO files generated by po4a 0.57
It looks like the recent po4a extracts table information
differently from the past version. This leads to many
fuzzy/untranslated strings in ja.po.
To clarify the further work, the generated PO files are
committed as a snapshot. Draft files are not updated in
this commit as I don't want to lose existing translations.