ebtables: Rename ebtables.8 to ebtables-legacy.8
The implementation of ebtables was changed to nftable-based
and the man page of the legacy implementation was renamed to
ebtables-legacy.8 in the recent ebtables release (2.0.11).
Recent distributions like Ubuntu 20.04 already contains
the newer implementation and both of ebtables-nft.8 and
ebtables-legacy.8 are available. Considering this,
it is confusing to provide ebtables.8.

This commit updates the version of ebtables to 2.0.11 and
renames JM deliverable ebtables.8 to ebtables-legacy.8.
Texts added in 2.0.11 are added as-is (without translating them).