LDP: Handle attribute table (on thread safety) in all PO files
The attribute tables were added to many pages in the recent LDP man-pages.
It looks better that these tables are translated consistently,
so this commit handles all appearances of them.

As of now, the only attribute described in these tables is "thread safety"
and attribute values are kind of labels, so they should be used as-is
in translated documents. If you have questions, I believe most questions
will be solved if you see attributes(7).