LDP: Fix translation_list
git2upd script assumes that a single blank line exists between
the section header of "Changes to be committed" and the next line
and a list of staged files, but the recent version of git does not
output such blank line (at least git 2.25.1 does not output the blank
line and git 2.17.1 outputs the blank line) and as a result git2upd
will pick up no staged files.

This commit recovers such changes in LDP 5.10.

What I did is to save the output of "git status original" to a file,
add a blank line before a list of staged files and then run git2upd.