LDP: Do not execute git2upd in 'make upgrade'
It turns out that git2upd does not work expectedly in the recent
version of git status. As a quick look at git2upd, it needs to be
fixed carefully. As a temporary workaround, git2upd and subsequent
commands in "make upgrade" are not executed automatically. They
need to be run manually.

The JM guide is updated accordingly.