coreutils: (Makefile) Copy generated pages to draft by default
Previously, when running 'make', manpages were generated in
help2man.ja directory but these files are not tracked by git.
It would be nice if we can see all changes you make by git.
This commit changes to call 'make instal' by default.

'setup' target is also added as a synonym for 'source' target.
LDP man-pages directory uses 'make setup' for preparation.
It is nice if we can use consistent target names as it is easy
to understand.

.PHONY targets are also declared now.