LDP: (capabilities.7) Fix typo reported in [JM:01235]
Note that the translation progress of capabilities.7 does not
satisfy the threshold for publishment, so for the latest version
(LDP 5.10) this commit fixes the typo reported in ja.po.

The previous version of the released page in "releases"
directory was also fixed directly. By doing so, the published
older version also contains the correct information.