dblatex: workaround for unexpexted indent in HTML
When man2html converts dblatex.1, it seems that man2html misunderstood
"DB2LaTeX" word as some roff directive if it is placed at the beginning
of line. As a result, the generated HTML contains <DD> HTML tag without
a closing tag </DD>. This leads to a situation that lines below this are
indented unexpexted. Apparently this is a bug of man2html we used,
but as a quick look it is not easy to be fixed.
A dirty workaround is to wrap lines before "これは、 DB2LaTeX ..."
to avoid starting a line with "DB2LaTeX".